% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_str_to_obj.R
\name{convert_str_to_obj}
\alias{convert_str_to_obj}
\title{Internal function used for testing}
\usage{
convert_str_to_obj(x)
}
\arguments{
\item{x}{string}
}
\description{
Internal function used for testing
}

