% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tdm_json_object.R
\name{convert_tdm_json_object}
\alias{convert_tdm_json_object}
\title{Safely convert the tdm object in JSON}
\usage{
convert_tdm_json_object(tdm)
}
\arguments{
\item{tdm}{tdm object}
}
\description{
Safely convert the tdm object in JSON
}

