% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nested_value.R
\name{get_nested_value}
\alias{get_nested_value}
\title{Get nested value from a list}
\usage{
get_nested_value(l = list(), address = "")
}
\arguments{
\item{l}{list}

\item{address}{address in list, nesting separated using . (similar to $ in R)}
}
\description{
Get nested value from a list
}

