% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2test-package.R
\docType{package}
\name{json2test}
\alias{json2test}
\title{json2test Package}
\description{
Functions for specifying tests and references in JSON format
}
\author{
Ron Keizer \email{ron@insight-rx.com}
}
