% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_test.R
\name{json_test}
\alias{json_test}
\title{#' Wrapper function to run tests and compare with reference}
\usage{
json_test(module = NULL, package = "insightrxr", tests = NULL,
  reference = NULL, equal = TRUE, delta = 0.03, run = NULL,
  skip = c(), ignore_keys = c("comment", "comments"))
}
\arguments{
\item{module}{linking to test object in tests.json}

\item{package}{package name from where to take tests.json object from}

\item{tests}{a vector of tests to run}

\item{reference}{list of reference values}

\item{equal}{should answers be equal? if not, specify allowed relative `delta`}

\item{delta}{relative allowed imprecision, default is 0.03. Only used when not specified in reference JSON.}

\item{run}{a specific test to run only, not check. The output object will be returned.}

\item{skip}{a vector of tests to skip (default is to run all)}

\item{ignore_keys}{ignore specific keys in reference JSON, e.g. to allow for comments}
}
\description{
#' Wrapper function to run tests and compare with reference
}

