% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_test_all.R
\name{json_test_all}
\alias{json_test_all}
\title{Run all JSON tests included with package}
\usage{
json_test_all(func = NULL, package = NULL, skip_func = c(), ...)
}
\arguments{
\item{func}{function to test, also refers to name of JSON file in test/reference folders}

\item{package}{name of package}

\item{skip_func}{skip this functions (vector of character)}

\item{...}{arguments to pass to `json_test` function}
}
\description{
Run all JSON tests included with package
}
