% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arg.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse arguments
Taken from opencpu, to remove dependencies for builds}
\usage{
parse_arg(x)
}
\arguments{
\item{x}{object}
}
\description{
Parse arguments
Taken from opencpu, to remove dependencies for builds
}
