% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json_test.R
\name{parse_json_test}
\alias{parse_json_test}
\title{Safely parse JSON to R object}
\usage{
parse_json_test(obj, parse_functions = list())
}
\arguments{
\item{obj}{object read from JSON}

\item{parse_functions}{list of function names, keys are the elements in the object}
}
\description{
Safely parse JSON to R object
}

