% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_test.R
\name{json_test}
\alias{json_test}
\title{Wrapper function to run tests and compare with reference}
\usage{
json_test(
  func = NULL,
  package = NULL,
  test_id = NULL,
  tests = NULL,
  reference = NULL,
  delta = NULL,
  run = NULL,
  skip = c(),
  list_as_args = TRUE,
  ignore_keys = c(),
  parse_functions = list(),
  overwrite = NULL,
  fail_if_not_exists = TRUE,
  use_testthat = FALSE
)
}
\arguments{
\item{func}{linking to test/reference JSON and function in package to test}

\item{package}{package name from where to take tests.json object from}

\item{test_id}{specifies test/reference JSON to use, i.e. /home/user/R/3.3/package_name/test/<test_id>.json. If not specified, assumes `func` is used as `test_id`.}

\item{tests}{a vector of tests to run}

\item{reference}{list of reference values}

\item{delta}{relative allowed imprecision, default is 0.03. Overrides value specified in reference JSON.}

\item{run}{a specific test to run only, not check. The output object will be returned.}

\item{skip}{a vector of tests to skip (default is to run all)}

\item{list_as_args}{should the list elements from the JSON be used as arguments to the function? TRUE by default. If FALSE, the list as a whole will be passed to the `args` argument of the function.}

\item{ignore_keys}{ignore specific keys in reference JSON, e.g. to allow for comments}

\item{parse_functions}{list of functions to parse specific JSON tests data before calling the function. This is sometimes useful due to the back-serialization from JSON to R object.}

\item{overwrite}{optional list specifying what keys of test JSON to manually overwrite with given values}

\item{fail_if_not_exists}{fail if folder with JSON does not exist?}

\item{use_testthat}{use testthat library}
}
\description{
Wrapper function to run tests and compare with reference
}
