% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arg.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse string, assuming JSON format
If arguments are not valid JSON, returns stop. Adapted from opencpu.}
\usage{
parse_arg(x)
}
\arguments{
\item{x}{object}
}
\description{
Parse string, assuming JSON format
If arguments are not valid JSON, returns stop. Adapted from opencpu.
}
